Jarvis-2025-master/requirements.txt
eel>=0.14.0
opencv-python>=4.5.0
numpy>=1.19.0
face-recognition>=1.3.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
pyaudio>=0.2.11 
requests
python-dotenv 